requirements.txt
Pillow==10.0.1
streamlit==1.27.2
pandas==2.1.1
matplotlib==3.8.0